requirements.txt
Pillow
GitPython
pytz
